docker-requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-rest-inherits-error-messages==1.0.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-spectacular==0.22.1
gevent==21.12.0
greenlet==0.4.13
gunicorn==20.1.0
hpy==0.0.3
idna==3.3
importlib-resources==5.8.0
inflection==0.5.1
jsonschema==4.6.2
packaging==21.3
PyMySQL==1.0.2
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
readline==6.2.4.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
